intervals for marketcap chart
