+++ src/ducks/Stablecoins/StablecoinsMarketCap/StablecoinsMarketCap.js
@@ -1,17 +1,18 @@
 import React, { useEffect, useMemo, useState } from 'react'
 import cx from 'classnames'
 import withSizes from 'react-sizes'
-import MarketCapHeader, {
-  MARKET_CAP_DAY_INTERVAL,
-  MarketcapIntervals
-} from './MarketCapHeader/MarketCapHeader'
+import StablecoinsHeader, {
+  StablecoinsIntervals
+} from './MarketCapHeader/StablecoinsHeader'
 import CheckingAssets from './CheckingAssets/CheckingAssets'
-import { convertToSeconds } from '../../dataHub/metrics/intervals'
 import Chart from '../../Chart'
 import { metricsToPlotCategories } from '../../Chart/Synchronizer'
 import {
+  getIntervalDates,
+  MARKET_CAP_DAY_INTERVAL,
   METRIC_SETTINGS_MAP,
   METRIC_TRANSFORMER,
+  STABLE_COINS_MARKETCAP_INTERVALS,
   STABLE_COINS_METRICS
 } from './utils'
 import { useTimeseries } from '../../Studio/timeseries/hooks'
@@ -19,16 +20,6 @@ import { useChartMetrics, useMetricColors } from './hooks'
 import { DesktopOnly, MobileOnly } from '../../../components/Responsive'
 import { mapSizesToProps } from '../../../utils/withSizes'
 import styles from './StablecoinsMarketCap.module.scss'
-
-export const getIntervalDates = interval => {
-  return {
-    from: new Date(
-      new Date().getTime() + -1 * convertToSeconds(interval.value)
-    ),
-    to: new Date(),
-    interval: '1h'
-  }
-}
 
 const CHART_HEIGHT = 400
 const CHART_PADDING_DESKTOP = {
@@ -79,11 +70,15 @@ const StablecoinsMarketCap = ({ isDesktop, className }) => {
 
   return (
     <div className={cx(styles.container, className)}>
-      <MarketCapHeader title='Stablecoins Market Cap'>
+      <StablecoinsHeader title='Stablecoins Market Cap'>
         <DesktopOnly>
-          <MarketcapIntervals interval={interval} setInterval={setInterval} />
+          <StablecoinsIntervals
+            interval={interval}
+            setInterval={setInterval}
+            intervals={STABLE_COINS_MARKETCAP_INTERVALS}
+          />
         </DesktopOnly>
-      </MarketCapHeader>
+      </StablecoinsHeader>
 
       <DesktopOnly>
         <CheckingAssets
@@ -110,7 +105,11 @@ const StablecoinsMarketCap = ({ isDesktop, className }) => {
       />
 
       <MobileOnly>
-        <MarketcapIntervals interval={interval} setInterval={setInterval} />
+        <StablecoinsIntervals
+          interval={interval}
+          setInterval={setInterval}
+          intervals={STABLE_COINS_MARKETCAP_INTERVALS}
+        />
         <CheckingAssets
           loadings={loadings}
           toggleDisabled={setDisabledAsset}
